Task 1: Improved login.test.ts, edited localStorage.setup.ts, and added existingUsers.ts
